requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
Django==2.0
django-haystack==2.8.1
idna==2.7
jieba==0.39
Markdown==2.6.11
mysqlclient==1.3.13
natgrid==0.2.1
pycparser==2.18
Pygments==2.2.0
PyMySQL==0.9.2
pytz==2018.5
six==1.11.0
Whoosh==2.7.4
